refactor: split frontmatter parsing + VerifyOutcome out of verify.py (#242)
Pure domain modules imported backward into verify.py (a 1,791-line
git/subprocess module): stories.py pulled read_frontmatter/status_of and
escalation.py pulled VerifyOutcome, dragging the whole git surface into
modules that have no need for it (assessment finding F-1).

- New src/bmad_loop/frontmatter.py holds read_frontmatter, status_of, and
  set_frontmatter_status — pure file/markdown parsing, zero git imports.
- VerifyOutcome (frozen dataclass, no git deps) moves to model.py.
- verify.py re-exports all moved names (explicit noqa: F401 on the
  re-export-only set_frontmatter_status, which the F401 autofix would
  otherwise delete) so every existing verify.<name> / from .verify import
  <name> call site and test stays green.
- stories.py and escalation.py import from the new homes.

Behavior-preserving move only — no logic edits.
